requirements-min.txt
fastapi==0.121.2
uvicorn==0.38.0
langchain==1.0.3
langchain-core==1.0.2
langgraph==1.0.2
langgraph-prebuilt==1.0.2
pydantic==2.12.3
httpx==0.28.1
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
beautifulsoup4==4.14.3
pandas==2.2.2
numpy==2.2.6
openpyxl==3.1.5
pypdf==6.4.1
docx2python==3.5.0
pillow==10.3.0
lxml==6.0.2
cozeloop==0.1.25
coze-coding-utils==0.2.4
